requirements.txt
Django~=5.1.5
djangorestframework~=3.15.2
pre_commit~=4.1.0
pytest-django~=4.9.0
redis~=5.2.1
celery[redis]~=5.4.0
pytest~=8.3.4
psycopg2-binary~=2.9.9